Substitute tags for holding the DOMElement itself
Using the tags in Subject gave us more knowledge of the issue, but was causing a bottleneck in the sanitiser. This was especially noticable when using large SVGs, such as the one in useDosTestTwo.svg.

Instead of using tags, I’ve now opted to store the afected elements (\DOMElement) directly within an array, within the Resolver class.

Doing this has allowed us to remove all tag checking and therefore the bottleneck that was being caused. It’s also allowed us to free up memory within Subject as once we have the affected elements in the array, we can remove the internal references.